requirements.txt
asgiref==3.6.0
Django==4.2
djangorestframework==3.14.0
exceptiongroup==1.1.1
iniconfig==2.0.0
packaging==23.1
pluggy==1.0.0
pytest==7.3.1
pytest-django==4.5.2
pytz==2023.3
sqlparse==0.4.4
tomli==2.0.1
